[PSDK-3] Incorporate JWT generation and client library into SDK (#5)
### What changed? Why?
Incorporates JWT generation based on API key, and the OpenAPI spec-based
client library into the SDK.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
Adds the ability to access the current user.