Use wallet model ID instead of generating new one (#14)
### What changed? Why?
Use the wallet model for the wallet ID rather than generating a new one.
This was causing a bug in Wallet import / export.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
Fix bug