Update Wallet::Data to be keyword initialized (#21)
### What changed? Why?
This makes the Coinbase::Wallet::Data object keyword initialized.

Since users have to persist this data they will need to rehydrate this.
Having the ordering matter makes it more likely for developers to= run
into errors (as I did when integrating the SDK).

This also adds a regression test for this PR:
https://github.com/coinbase/coinbase-sdk-ruby/pull/20

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->