Support requesting faucet funds for address (#24)
### What changed? Why?

This adds support for requesting faucet funds for a given address. This
will be rate limited by the address via the smart contract that backs
the faucet and via the API key (1 request every 24 hours).

This may be able to be loosened for base-sepolia via the Platform API,
but those are the current rate limit parameters.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This adds a new functionality to an unreleased version. This only
impacts the address model.