Remove timestamp from generated doc footer (#34)
### What changed? Why?
This removes the timestamp from the docs generation footer template so
that the docs don't change unless the content actually changes.

This also updates the generated docs for latest `v0.0.3` changes

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->

Docs changes only.