Consolidate asset and balance handling
This consolidates the transforms around assets and balances to
their respective classes, and moves them out of wallet and address.

This also adds testing and fixes the handling for fetching balances
of `weth` using `get_address_balance` and `get_wallet_balance` which
was not handled properly.

The `list_address_balances` and `list_wallet_balances` handling properly
handled `weth`.