Refactor method names and where asset, balance, and import logic lives (#41)
### What changed? Why?

This renames some methods to be more idiomatic, e.g.
`address.list_balances` => `address.balances`.

This also moves the responsibility of some logic into the class that
best represents the domain of that behavior.

* Wallet Import
* `user.import_wallet` => `Coinbase::Wallet.import` (the helper still
exists, but logic belongs to Wallet).
* Balance Map
  * `Coinbase.to_balance_map` => `Coinbase::BalanceMap.from_balances`
* Asset denomination handling
* Logic existed at the wallet and address layer now belongs to `Asset`.
* Coinbase::Balance handles parsing client balances and converting them
to the write asset using the asset class methods.
* `address.list_transfer_ids` => `address.transfers`.
* Change ID methods to not stutter
  * `user.user_id` => `user.id`
  * `wallet.wallet_id` => `wallet.id`
  * `address.address_id` => `address.id`


TODO:
* Move asset and amount manipulation to the backend. Supporting gwei and
wei and taking an argument to indicate if we're passing `whole` or
`atomic` units will allow us to not have to do any of this logic client
side.
* Move wallet creation logic out of `Wallet.new` so we can leverage
wallets in a state where we do not need a seed to be imported.
* This enables us to add user.wallets or Coinbase::Wallet.list that
returns a set of wallets

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->

This contains some breaking changes from v0.0.3.

---------

Co-authored-by: Yuga Cohler <yuga.cohler@coinbase.com>