Support configuring client scheme via API URL (#46)
### What changed? Why?
This supports configuring the underlying OpenAPI REST HTTP client scheme
(i.e. `http` or `https`) via the already configured API URL.

This will enable us to configure the SDK to run against the backend
locally in E2E regression tests.

Note: The OpenAPI client library only strips the `https` scheme from the
`host`, but by default the host should omit the scheme.

In order to support this we explicitly set the `host` and optionally
include the `port` only if specified.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->

If developers are configuring API URL's without a protocol then the
scheme could