[chore] Refactor wallet creation
This moves wallet creation into a wallet class method and splits
out default address creation from the initializer. This makes it
so that the initializer is just intitializing and fetching resources
and not actually mutating things.

When we create the wallet initially we will create the default
address, but we don't need that code path for every single initialize
call.

This also adds dynamic network support as the backend is keeping the
source of truth for supported networks.

This also updates the tests so that we're loading / importing wallets
for most scenarios.

[PSDK-86] Fix create_wallet for ruby 2.7

This makes it so that the handling for passing through the
create_wallet arguments from `user.create_wallet` to `Wallet.create`
use syntax compatible with ruby 2.7