[chore] Refactor wallet creation (#50)
### What changed? Why?
This moves wallet creation into a wallet class method and splits out
default address creation from the initializer. This makes it so that the
initializer is just intitializing and fetching resources and not
actually mutating things.

When we create the wallet initially we will create the default address,
but we don't need that code path for every single initialize call.

This also adds dynamic network support as the backend is keeping the
source of truth for supported networks.

This also updates the tests so that we're loading / importing wallets
for most scenarios.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->

This touches the `user.create_wallet` flow, so wallet creation could
break.