[chore] Run unit tests against various ruby versions (#49)
### What changed? Why?
This runs our unit tests against the ruby versions going back to the
latest one we support (`2.7`).

This explicitly adds `bigdecimal` as a runtime dependency as it is no
longer bundled with ruby as of 3.4


#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->