[PSDK-86] Enable creation of wallets backed by server signers (#48)
### What changed? Why?
This enables developers to create wallets that are backed by server
signers v.s. managing the seeds themselves.

When calling `user.create_wallet(server_signer: true)` the wallet will
be created and associated with the project's default server signer (if
one exists). It will return an error if a server signer does not exist
for that project.

### Default address creation

Rather than add further cases to the wallet initialization this removes
the side-effects from that method in favor of a `Wallet.create` class
method.

This code path will only be used for creation of wallets, where we will
additionally create a default address for wallets that are not backed by
server signers.

### Network IDs

This also adds support for specifying the `network_id` which the backend
will properly validate based on supported networks.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This touches the `user.create_wallet` flow