[chore] Add test coverage reporting (#56)
### What changed? Why?
This adds basic simplecov test coverage reporting that is useful for
testing unit test coverage locally.

This can be pushed to a github action step and enforced if needed
separately as there weren't any official reporting actions that I saw.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This only impacts tests