[chore] Add test coverage and refactor `Address#transfer` (#57)
### What changed? Why?

This adds test coverage around the Address model, specifically the
`#transfer` and `#transfers` methods.

This also refactors `Address#transfer` to make it a bit easier to parse
in anticipation for transferring using server signer and supporting
trading.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This could cause issues with the Transfer feature