[chore] Move simplecov to gemspec dev dependency (#58)
This moves the simplecov gemspec as a development dependency so that
tests can be ran using simplecov in a development context, as well as
have our E2E tests in CI run using simplecov.

### What changed? Why?


#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->