[PSDK-138] Add trade support (#60)
### What changed? Why?

This adds trade support for address via:

```
address.trade(0.1, :eth, :usdc)
```

This utilizes the new trade APIs to construct a trade resource and
server-side unsigned transaction, sign it within the SDK, and then
broadcast it back using the server.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->

None, merging into a feature branch