[chore] Increase wait time in transfer (#64)
### What changed? Why?

Now that server-signer could also sign transactions, wait time of 10s is
low in certain scenarios. Signer polling frequency is 10s, so we want
the sdk timeout to be longer than that.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->