[PSDK-210] Handle approve transactions in the Trade flow (#65)
### What changed? Why?
This adds handling for Approve Transactions in the Trade flow.

When trading between two ERC20 assets we also need to construct,
and sign the approve transaction alongside the actual trade transaction.

The backend will ensure this transaction gets broadcast first.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This is a new trade feature going into a feature branch.