Add Transaction class (#73)
### What changed? Why?
This adds the Coinbase::Transaction class that will be used by trading,
staking, and transfers.

We may want to consider the `sign` function to actually update the
`signed_payload` at some point

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This class is currently unused so should have no impact