Migrate transfers to use new transaction class (#74)
### What changed? Why?
This migrates the transfers to use the newly added transaction class.

This will unify all of our different interactions transfer/trade/stake
to interact with the same transaction object and make flows such as
signing, etc... normalized.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->