Feature/trading (#72)
### What changed? Why?

This adds trade support for address via:

```
address.trade(0.1, :eth, :usdc)
```

This utilizes the new trade APIs to construct a trade resource,
server-side unsigned transaction(s), sign the transaction(s) within the
SDK, and then broadcast it back using the server.

When performing ERC20 to ERC20 trades there are two transactions that
need to be signed and broadcast, the first being an approve transaction
and the latter being the transaction that will perform the trade.

This is only supported on `base-mainnet` currently as there is limited
support for swaps on testnet (but this may be supported down the road).

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This is a new feature

---------

Co-authored-by: Jayasudha Jayakumaran <121061531+jazz-cb@users.noreply.github.com>