Support transaction links across networks (#77)
### What changed? Why?
This adds support for transaction links to block explorers across all of
the networks that the Coinbase Ruby SDK supports.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
Transaction links could be accidentally omitted