Use progress format when running locally (#83)
### What changed? Why?

Hey all! I realized when doing local development the documentation
format of rspec can get kinda noisy. It hid away some warnings I was
seeing. I believe we could benefit from using the progress formatter
instead where successes are hidden as a `.` and warnings are easier to
spot.

#### Qualified Impact
This has no impact on the gem itself, only development of it.