Enable ethereum networks for address path prefix (#84)
### What changed? Why?
This handles both ethereum and base networks for address path prefix.
This will be pushed to the backend so we don't need to know what
protocol each network is on the client-side.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->