Lazily load wallet addresses
This makes it so that when we are loading wallet's we do not
need to load all of the addresses up front.

The N+1 query when loading wallets, makes listing wallets a
prohibitively expensive and slow request.

This makes it so that we can list all of the wallets in a set of
pagination and enumerable requests.

Then for each loaded wallet, when you attempt to interact with
the addresses we will lazily fetch them for you.

This also handles updating the keys for these cached addresses
when a seed is set.