Support listing and fetching wallets (#91)
### What changed? Why?

This adds support for easily listing and fetching wallets without the
overhead of the N+1 of fetching all of the addresses for all of the
wallets we've fetched.

#### Listing wallets
```ruby
Coinbase::Wallet.list
```

```ruby
Coinbase.default_user.wallets
```

#### Fetching a wallet

Fetching a wallet returns a wallet object that is not in a signing
state. If using the server signer then this wallet can be used for
signing purposes however! If not using the server signer this wallet and
be loaded with a seed and then used for signing.

```ruby
Coinbase::Wallet.fetch(wallet_id)
```

```ruby
Coinbase.default_user.get_wallet(wallet_id)
```


#### Qualified Impact
This changes the `user.wallets` behavior from something that returns an
array of 2 elements, the first being an array of wallets and the second
being a page token, into a proper enumerable response that the user can
use to lazily fetch all wallets if needed.