Consolidate address balance and faucet endpoints (#96)
### What changed? Why?
This consolidates the address balance and faucet endpoints to use the
non-wallet scoped ones.

This will enable us to consolidate our APIs eventually


#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->

This is a cleanup task and should not impact any behavior.