Stone/v0.0.9 cleanup (#99)
### What changed? Why?
This cleanups the v0.0.9 release by removing an unused `jimson`
dependency, fixing a typo in the README, and rebuilding the docs for
v0.0.9.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->