Migrate staking context to use the balance objects instead of in-line balances (#101)
### What changed? Why?

This migrates the staking context to use the balance objects instead
of the in-line balances.

This also utilizes the fact that the API returns `"0"` for balances
rather than the empty string so we can remove the unnecessary
client-side handling.

```
> address = Coinbase::ExternalAddress.new(:ethereum_holesky, '0x87Bf57c3d7B211a100ee4d00dee08435130A62fA')
=> Coinbase::Address{id: '0x87Bf57c3d7B211a100ee4d00dee08435130A62fA', network_id: 'ethereum_holesky'}
> address.staking_balances(:eth)
=> {:stakeable_balance=>0.1655172821856387445e1, :unstakeable_balance=>0.2108042377507975952e1, :claimable_balance=>0.113465e-12}
```

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->