[chore] Decode raw transaction from signed payload when possible.
This makes it so that we populate a signed raw transaction when
it is returned from the backend with a signed payload.

This ensures that when we have actually signed a transaction, and
you fetch it from the backend it returns `transaction.signed?` as
true instead of false.

Previously it would only return true when we've signed the transaction
in band.