Refactor WalletAddress to use Trade helpers
This refactors the WalletAddress class to use the Trade class helpers
for listing trades, creating trades, and broadcasting trades.

This reduces the scope of responsibility for the WalletAddress class
in the trade flow, so it does not need to know details about what the
different transaction types are or how to broadcast the various
signed payloads.