Simplify Trade, Transfer, and Transaction responsibilities (#104)
### What changed? Why?

This refactors the `Transaction`, `Trade`, and `Transfer` classes to be
responsible for more of their own business logic, as well as to get us
closer to supporting the `ExternalAddress` paradigm for
trades/transfers/staking.

#### Transaction
This refactors the transaction payload to properly handle being decoded
from the signed payload. This ensures that when we have actually signed
a transaction, and you fetch it from the backend it returns
`transaction.signed?` as true instead of false. Previously it would only
return true when we've signed the transaction in band.

#### Refactoring Trade & Transfer classes

This refactors the Trade & Transfer classes to expose a `.list`,
`.create`, and `#broadcast` methods so that they can be responsible for
their domain. For example previously the `WalletAddress` had to have
context into whether a trade needs an approve transaction signed, plus
how to broadcast the payload contextually.

Now the `WalletAddress` is only responsible for
1. Creating Trade/Transfer
2. Signing Trade/Transfer transactions
3. Broadcasting Trade/Transfer

Previously it had to know details about the inner workings of a trade,
e.g. what an approve transaction is, etc...

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->