Enable all new rubocops and add rubocop-rspec
This enables all new rubocops, adds rubocop-rspec, and removes the
high limitations of certain `max` configured values.

Renaming the E2E tests is difficult because we call rspec spec/e2e/end_to_end.rb
across multiple versions, so we'd need to add handling to support that or
duplicate the file for a few versions, neither of which feels better than just
disabling the linting rule.