Enable all new rubocop linters and configure rubocop-rspec linting (#132)
### What changed? Why?
This updates our linting rules to add new cops as well as rubocop-rspec.
This also addresses most of the violations.

Note: Renaming the E2E tests is difficult because we call `rspec
spec/e2e/end_to_end.rb` across multiple versions, so we'd need to add
handling to support that or duplicate the file for a few versions,
neither of which feels better than just disabling the linting rule.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This mostly impacts minor linting related things, but touches some
flows:
* Server Signer Status
* Wallet Data / Importing