[chore] Fix MPC wallet address initailization (#154)
### What changed? Why?
Previously when creating MPC wallet's we were accidentally appending 2
identical wallet addresses to the cached addresses.

This makes it so that when we're initially creating the wallet, we just
do a full model reload instead of using the local cache.


#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->