Refactor Wallet class to use delegators (#136)
### What changed? Why?
This refactors the Wallet class to use explicit delegators for
forwarding to the default address so we do not need to redefine
arguments in two places and potentially have them become out of sync.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->