Pass address index when registering address (#150)
### What changed? Why?
This makes it so that the SDK passes the address index when deriving
and registering addresses so that we can have a strong guarantee
that the addresses are listed in address index order.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->