Inject correlation context headers (#158)
### What changed? Why?
This injects the correlation context as request headers including the
SDK version and language. This will better help assess what client SDK
versions our developers are using so we can proactively reach out to
developers on older versions.

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->