Support deploying smart contract tokens (ERC20)
This adds support for deploying Smart Contract tokens, specifically
ERC20.

This adds a `deploy_token` function to wallet and wallet addresses,
that enable us to deploy an ERC20 token smart contract with the
specified token name, symbol, and total supply.

This token contract will be interactable and manageable with the
deploying MPC / dev-managed wallet.