Support contract invocations without ABI or amount
This adds fixes for invoking contracts without an ABI, which is
now supported for ERC20, ERC721, and ERC1155 standard contracts.

This also fixes handling for contract invocations that omitted a
payable amount, as we were receiving a backend error when passing
through the `null` value, v.s. omitting the key.