chore: Bump OpenAPI client (#216)
### What changed? Why?

This bumps the OpenAPI client version and fixes the client API names

* `Wallet Stake` => `MPC Wallet Stake`
* `Validators` -> `Stake`

#### Qualified Impact
<!-- Please evaluate what components could be affected and what the
impact would be if there was an
error. How would this error be resolved, e.g. rollback a deploy, push a
new fix, disable a feature
flag, etc... -->
This should have no impact and should be a no-op