Bug fix for non-checksummed address (#236)
### What changed? Why?
There was a bug in the from_model function in asset.py. On the backend
side, any asset id is checksummed before retrieving its details, and
when it returns a response to the SDK, its asset id is also checksummed.
When a user enters a non-checksummed address, they will get an error
since those two asset ids are not equivalent.

This change will allow users to use non-checksummed addresses without
getting erroneous errors in their workflow.

Before:
```
>>> wallet.balance('0x8309fbdf021edf768dc13195741940ba544dea98')

ArgumentError: Unsupported asset ID: 0x8309fbdf021edf768dc13195741940ba544dea98
```

After:
```
>>> wallet.balance('0x8309fbdf021edf768dc13195741940ba544dea98')

0.0
```

#### Qualified Impact
This change should not be a breaking change for the SDK. The same logic
of the check for asset ids exists, but with the normalization of cases,
it prevents the bug of non-checksummed addresses and checksummed
addresses being considered as unequal.