chore: bump version to v0.16.0 (#237)
What changed? Why?
The package versions and changelog were updated to reflect the upcoming
release of SDK version v0.16.0.

Qualified Impact
The impact of this change will update the SDK to the latest version
which includes one bug fix. In the case of any errors or failures, the
best approach would be to rollback to v0.14.0 and re-release a new
version after fixing any errors.