V0.16.0 (#238)
### What changed? Why?
A bug fix for non-checksummed addresses was added in this release. The
change was made to prevent users from encountering an "unsupported
asset" error when using non-checksummed addresses for assets.

#### Qualified Impact
This component mainly affects the assets class, where the asset id gets
normalized. If there are any issues, it should be safe to revert back to
v0.14.0 and then fix any errors before re-releasing.